Struct maidsafe_client::client::response_getter::ResponseGetter

pub struct ResponseGetter {
    // some fields omitted
}

ResponseGetter is a lazy evaluated response getter. It will fetch either from local cache or wait for the CallbackInterface to notify it of the incoming response from the network.

Methods

impl ResponseGetter

fn new(notifier: ResponseNotifier, cb_interface: Arc<Mutex<CallbackInterface>>, msg_id: Option<MessageId>, name: Option<NameType>) -> ResponseGetter

Create a new instance of ResponseGetter

fn get(&mut self) -> Result<Vec<u8>, ResponseError>

Get either from local cache or (if not available there) get it when it comes from the network as informed by CallbackInterface. This is blocking.