Struct maidsafe_client::client::Client

pub struct Client {
    // some fields omitted
}

The main self-authentication client instance that will interface all the request from high level API's to the actual routing layer and manage all interactions with it. This is essentially a non-blocking Client with upper layers having an option to either block and wait on the returned ResponseGetter for receiving network response or spawn a new thread. The Client itself is however well equipped for parallel and non-blocking PUTs and GETS.

Methods

impl Client

fn create_account(keyword: &String, pin: u32, password_str: &String) -> Result<Client, IoError>

This is one of the two Gateway functions to the Maidsafe network, the other being the log_in. This will help create a fresh account for the user in the SAFE-network.

fn log_in(keyword: &String, pin: u32, password_str: &String) -> Result<Client, IoError>

This is one of the two Gateway functions to the Maidsafe network, the other being the create_account. This will help log into an already created account for the user in the SAFE-network.

fn set_root_directory_id(&mut self, root_dir_id: NameType) -> Result<(), IoError>

Create an entry for the Root Directory ID for the user into the session packet, encrypt and store it. It will be retireved when the user logs into his account. Root directory ID is necessary to fetch all of user's data as all further data is encoded as meta-information into the Root Directory or one of its subdirectories.

fn get_root_directory_id(&self) -> Option<&NameType>

Get Root Directory ID if available in session packet used for current login

fn hybrid_encrypt(&self, data_to_encrypt: &[u8], nonce_opt: Option<Nonce>) -> Result<Vec<u8>, SymmetricCipherError>

Combined Asymmectric and Symmetric encryption. The data is encrypted using random Key and IV with AES-symmetric encryption. Random IV ensures that same plain text produces different cipher-texts for each fresh symmetric encryption. The Key and IV are then asymmetrically enrypted using Public-MAID and the whole thing is then serialised into a single Vec.

fn hybrid_decrypt(&self, data_to_decrypt: &[u8], nonce_opt: Option<Nonce>) -> Option<Vec<u8>>

Reverse of hybrid_encrypt. Refer hybrid_encrypt.

fn get_owner(&self) -> NameType

Owner of this client is the name of Public-MAID

fn put<T>(&mut self, sendable: T) -> Result<ResponseGetter, IoError> where T: Sendable

Put data onto the network. This is non-blocking.

fn get(&mut self, tag_id: u64, name: NameType) -> Result<ResponseGetter, IoError>

Get data from the network. This is non-blocking. Additionally this incorporates a mechanism of local caching. So if data already exists in the local cache, it is immediately returned via ResponseGetter and no networking penatly is payed. The functionality is abstracted to the user and is baked entirely into ResponseGetter.