Enum maidsafe_client::errors::ClientError

pub enum ClientError {
    StructuredDataHeaderSizeProhibitive,
    UnsuccessfulEncodeDecode,
    AsymmetricDecipherFailure,
    SymmetricDecipherFailure,
    RoutingFailure(Error),
    ReceivedUnexpectedData,
    VersionCacheMiss,
    RoutingMessageCacheMiss,
    NetworkOperationFailure(ResponseError),
    RootDirectoryAlreadyExists,
    GenericIoError(Error),
}

Client Errors

Variants

StructuredDataHeaderSizeProhibitive

StructuredData has no space available to fit in any user data inside it.

UnsuccessfulEncodeDecode

Could not Serialise or Deserialise

AsymmetricDecipherFailure

Asymmetric Key Decryption Failed

SymmetricDecipherFailure

Symmetric Key Decryption Failed

RoutingFailure

Routing GET, PUT, POST, DELETE Immediate Failure

ReceivedUnexpectedData

ReceivedUnexpectedData

VersionCacheMiss

No such data found in local version cache

RoutingMessageCacheMiss

No such data found in routing-filled cache

NetworkOperationFailure

Network operation failed

RootDirectoryAlreadyExists

Cannot overwrite a root directory if it already exists

GenericIoError

Generic I/O Error