Struct maidsafe_client::nfs::io::reader::Reader

pub struct Reader {
    // some fields omitted
}

Reader is used to read contents of a File. It can read in chunks if the file happens to be very large

Methods

impl Reader

fn new(file: File, client: Arc<Mutex<Client>>) -> Reader

Create a new instance of Reader

fn size(&self) -> u64

Returns the total size of the file/blob

fn read(&mut self, position: u64, length: u64) -> Result<Vec<u8>, String>

Read data from file/blob