Struct maidsafe_client::nfs::io::writer::Writer

pub struct Writer {
    // some fields omitted
}

Writer is used to write contents to a File and especially in chunks if the file happens to be too large

Methods

impl Writer

fn new(directory: DirectoryListing, file: File, client: Arc<Mutex<Client>>, mode: Mode) -> Writer

Create new instance of Writer

fn write(&mut self, data: &[u8], position: u64)

Data of a file/blob can be written in smaller chunks

fn close(self) -> Result<(), String>

close is invoked only after alll the data is completely written The file/blob is saved only when the close is invoked.