Struct maidsafe_client::nfs::rest::Container

pub struct Container {
    // some fields omitted
}

Container Repersents a Directory. Container can have its own metadata, sub-containers and files

Methods

impl Container

fn authorise(client: Arc<Mutex<Client>>, dir_id: Option<[u8; 64]>) -> Result<Container, String>

Authorises the directory access and returns the Container, if authorisation is successful. Operations can be performed only after the authorisation is successful.

fn create(&mut self, name: String) -> Result<(), String>

Creates a Container

fn get_created_time(&self) -> Tm

Returns the Created time of the container

fn get_id(&self) -> [u8; 64]

Return the unique id of the container

fn get_metadata(&self) -> Option<String>

Returns the user metadata saved as String. None can be passed to clear the metadata

fn get_name(&self) -> &String

Returns the name of the container

fn get_blobs(&self) -> Vec<Blob>

Returns the list of Blobs in the container

fn get_blob(&self, name: String, version: Option<[u8; 64]>) -> Result<Blob, String>

Returns a Blob from the container

fn get_containers(&self) -> Vec<ContainerInfo>

Returns the list of child containers

fn get_versions(&mut self) -> Result<Vec<[u8; 64]>, String>

Retrieves Versions for the container

fn get_container_versions(&mut self, container_id: [u8; 64]) -> Result<Vec<[u8; 64]>, String>

Retrieves Versions for the container being referred by the container_id

fn get_container(&mut self, name: String, version: Option<[u8; 64]>) -> Result<Container, String>

Fetches the latest version of the child container. Can fetch a specific version of the Container by passing the corresponding VersionId.

fn delete_container(&mut self, name: String) -> Result<(), String>

Deletes the child container

fn create_blob(&mut self, name: String, metadata: Option<String>) -> Result<Writer, String>

Creates a Blob within the container Returns a Writter object The content of the blob is written using the writter. The blob is created only after the writter.close() is invoked

fn update_blob_content(&mut self, blob: &Blob, data: &[u8]) -> Result<(), String>

Updates the blob content. Writes the complete data and updates the Blob

fn get_blob_writer(&mut self, blob: &Blob) -> Result<Writer, String>

Return a writter object for the Blob, through which the content of the blob can be updated This is useful while handling larger files, to enable writting content in parts

fn get_blob_content(&mut self, blob: &Blob) -> Result<Vec<u8>, String>

Reads the content of the blob and returns the complete content

fn get_blob_reader(&mut self, blob: &Blob) -> Result<Reader, String>

Returns a reader for the blob Using a Reader helps in handling large file contents and also fetch data in a specific range

fn get_blob_versions(&mut self, name: String) -> Result<Vec<[u8; 64]>, String>

Returns the list of versions_id for the blob

fn update_blob_metadata(&mut self, name: String, metadata: Option<String>) -> Result<(), String>

Update the metadata of the Blob in the container

fn delete_blob(&mut self, name: String) -> Result<(), String>

Delete blob from the container

fn copy_blob(&mut self, blob_name: String, to_container: [u8; 64]) -> Result<(), String>

Copies the latest blob version from the container to the specified destination container