Struct maidsafe_client::nfs::rest::ContainerInfo

pub struct ContainerInfo {
    // some fields omitted
}

Wrapper over DirectoryInfo to present Rest-friendly name to the Restful interface users

Methods

impl ContainerInfo

fn get_id(&self) -> [u8; 64]

Get Container ID. This is the directory ID which is unique for every directory and is the only way to retrieve that directory (DirectoryListing) from the network

fn get_name(&self) -> &String

Get the name of the Container

fn get_created_time(&self) -> Tm

Get the creation time for this Container

fn convert_to_directory_info(&self) -> DirectoryInfo

Convert the ContainerInfo to the format of DirectoryInfo that lower levels understand and operate on

fn convert_from_directory_info(info: DirectoryInfo) -> ContainerInfo

Convert from the format of DirectoryInfo that the lower levels understand to the rest friendly ContainerInfo